Phase 3c.2 (#100): extract mapping_migrations.py from mapping_service
mapping_service.py was 4017 LOC after Phase 3c.1. This PR extracts the
schema DDL, version constants, backup helper, _init_db connection
factory, and the v2/v3/v4 migration functions into a new module.

What moved to scripts/web/services/mapping_migrations.py:
  - Constants: _SCHEMA_VERSION, _BACKUP_RETENTION, _SCHEMA_SQL
  - Helpers: _backup_db
  - Connection factory: _init_db
  - Migration functions: _migrate_v2_to_v3, _migrate_v3_to_v4

Re-exports for backward compat:
  mapping_service re-exports _init_db, _backup_db, _SCHEMA_VERSION,
  _BACKUP_RETENTION, _SCHEMA_SQL, _migrate_v2_to_v3, _migrate_v3_to_v4
  at module top-level. Reasoning: _init_db is the SQLite connection
  factory used by every query in mapping_service (~16 internal call
  sites) and by ~100 test sites in tests/test_mapping_service.py +
  tests/test_archive_*.py. A clean break would have churned 100+
  imports across the test suite for no behavioural benefit. The
  re-export is one line; new code SHOULD import from
  services.mapping_migrations directly. The same is NOT done for
  the queue functions (Phase 3c.1) — those have a much smaller
  caller surface and benefit from the clean break.

Stayed in mapping_service.py (runtime hot-path, shared with indexer):
  - _TRIP_GAP_MINUTES_DEFAULT, _MERGE_MAX_ITERATIONS
  - _merge_adjacent_trips_for, _merge_all_adjacent_trip_pairs
  - _haversine_km

Dependency direction is one-way:
  mapping_migrations does NOT import from mapping_service at module
  load time. The v2->v3 and v9 migration code lazy-imports
  _merge_all_adjacent_trip_pairs, _TRIP_GAP_MINUTES_DEFAULT, and
  _haversine_km from mapping_service inside the function bodies
  (those helpers stay where they are because they're hot-path
  dependencies of _index_video and shouldn't flicker between
  modules every time the migrations module loads).

mapping_service.py: 4017 -> ~3460 LOC (-557).
mapping_migrations.py: ~530 LOC.

Removed unused 'shutil' import from mapping_service.py (was only
used by _backup_db, which moved out).

Other cleanup:
  - .github/copilot-instructions.md updated with the new module
    location + the rule that new schema/migration code SHOULD import
    from services.mapping_migrations directly.

Tests: 1049 pass + 3 skipped (same totals as pre-PR; no test imports
needed updating thanks to the re-exports). Verified _init_db works
both via the new module and via the re-export shim, and that the
re-exported objects are the same Python objects (not duplicates).

Refs #100 (Phase 3c).
