+++ scripts/web/services/mapping_service.py
@@ -14,7 +14,6 @@ import json
 import logging
 import math
 import os
-import shutil
 import sqlite3
 import subprocess
 import threading
@@ -25,6 +24,25 @@ from enum import Enum
 from typing import Any, Callable, Dict, Generator, List, Optional, Tuple
 
 logger = logging.getLogger(__name__)
+
+
+# ---------------------------------------------------------------------------
+# Phase 3c.2 (#100): re-export schema/migration symbols.
+# ``_init_db`` is the connection factory used by every query function in
+# this module and by ~100 test sites. Re-exporting it (along with
+# ``_backup_db``, ``_SCHEMA_VERSION``, and ``_BACKUP_RETENTION``) keeps all
+# existing call sites working unchanged after the move.
+# New code should import from ``services.mapping_migrations`` directly.
+# ---------------------------------------------------------------------------
+from services.mapping_migrations import (  # noqa: E402,F401
+    _BACKUP_RETENTION,
+    _SCHEMA_SQL,
+    _SCHEMA_VERSION,
+    _backup_db,
+    _init_db,
+    _migrate_v2_to_v3,
+    _migrate_v3_to_v4,
+)
 
 
 # ---------------------------------------------------------------------------
@@ -140,517 +158,16 @@ def _with_db_retry(fn: Callable) -> Callable:
 
 
 # ---------------------------------------------------------------------------
-# Database Schema & Management
+# Database Schema & Migrations (Phase 3c.2 — moved to services.mapping_migrations)
 # ---------------------------------------------------------------------------
-
-_SCHEMA_VERSION = 10
-_BACKUP_RETENTION = 3  # Keep this many migration backups before pruning oldest
-
-_SCHEMA_SQL = """
-CREATE TABLE IF NOT EXISTS schema_version (
-    version INTEGER PRIMARY KEY
-);
-
-CREATE TABLE IF NOT EXISTS trips (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    start_time TEXT NOT NULL,
-    end_time TEXT,
-    start_lat REAL,
-    start_lon REAL,
-    end_lat REAL,
-    end_lon REAL,
-    distance_km REAL DEFAULT 0.0,
-    duration_seconds INTEGER DEFAULT 0,
-    source_folder TEXT,
-    indexed_at TEXT
-);
-
-CREATE TABLE IF NOT EXISTS waypoints (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    trip_id INTEGER REFERENCES trips(id) ON DELETE CASCADE,
-    timestamp TEXT NOT NULL,
-    lat REAL NOT NULL,
-    lon REAL NOT NULL,
-    heading REAL,
-    speed_mps REAL,
-    acceleration_x REAL,
-    acceleration_y REAL,
-    acceleration_z REAL,
-    gear TEXT,
-    autopilot_state TEXT,
-    steering_angle REAL,
-    brake_applied INTEGER DEFAULT 0,
-    blinker_on_left INTEGER DEFAULT 0,
-    blinker_on_right INTEGER DEFAULT 0,
-    video_path TEXT,
-    frame_offset INTEGER
-);
-
-CREATE TABLE IF NOT EXISTS detected_events (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    trip_id INTEGER REFERENCES trips(id) ON DELETE CASCADE,
-    timestamp TEXT NOT NULL,
-    lat REAL,
-    lon REAL,
-    event_type TEXT NOT NULL,
-    severity TEXT DEFAULT 'info',
-    description TEXT,
-    video_path TEXT,
-    frame_offset INTEGER,
-    metadata TEXT
-);
-
-CREATE TABLE IF NOT EXISTS indexed_files (
-    file_path TEXT PRIMARY KEY,
-    file_size INTEGER,
-    file_mtime REAL,
-    indexed_at TEXT,
-    waypoint_count INTEGER DEFAULT 0,
-    event_count INTEGER DEFAULT 0
-);
-
--- Persistent indexing work queue. One row per pending clip, keyed by
--- canonical_key (RecentClips/ArchivedClips dedup, Saved/Sentry events
--- disambiguated by event folder — see canonical_key()). The single
--- worker thread (services.indexing_worker) drains this; producers
--- (file watcher, archive job, manual button, boot catch-up) just
--- INSERT rows. claimed_by/claimed_at let the worker take an exclusive
--- claim atomically; stale claims (>30 min) are auto-released so a
--- crashed worker can't permanently lock a row.
-CREATE TABLE IF NOT EXISTS indexing_queue (
-    canonical_key TEXT PRIMARY KEY,
-    file_path TEXT NOT NULL,
-    priority INTEGER NOT NULL DEFAULT 50,
-    enqueued_at REAL NOT NULL,
-    next_attempt_at REAL NOT NULL DEFAULT 0,
-    attempts INTEGER NOT NULL DEFAULT 0,
-    last_error TEXT,
-    claimed_by TEXT,
-    claimed_at REAL,
-    source TEXT
-);
-
-CREATE INDEX IF NOT EXISTS idx_waypoints_trip ON waypoints(trip_id);
-CREATE INDEX IF NOT EXISTS idx_waypoints_coords ON waypoints(lat, lon);
-CREATE INDEX IF NOT EXISTS idx_waypoints_timestamp ON waypoints(timestamp);
-CREATE INDEX IF NOT EXISTS idx_waypoints_video_path ON waypoints(video_path);
--- Covering index for query_trips' video_count subquery: lets SQLite count
--- DISTINCT video_path per trip without touching the main waypoints table.
--- Without this, /api/trips fans out to 1 + 2N queries (where N = page size)
--- and visibly stalls the map page on databases with thousands of waypoints.
-CREATE INDEX IF NOT EXISTS idx_waypoints_trip_video
-    ON waypoints(trip_id, video_path);
--- Day-based aggregate (/api/days) and per-day route lookup
--- (/api/day/<date>/routes) both filter by substr(start_time,1,10).
--- v7: ``idx_trips_start_time`` keeps a sortable-text scan available
--- for callers that filter by ``start_time >= ?`` (e.g. /api/trips with
--- date_from/date_to).
-CREATE INDEX IF NOT EXISTS idx_trips_start_time ON trips(start_time);
-CREATE INDEX IF NOT EXISTS idx_events_trip ON detected_events(trip_id);
-CREATE INDEX IF NOT EXISTS idx_events_coords ON detected_events(lat, lon);
-CREATE INDEX IF NOT EXISTS idx_events_type ON detected_events(event_type);
-CREATE INDEX IF NOT EXISTS idx_events_timestamp ON detected_events(timestamp);
--- v8: expression indexes on substr(<ts>, 1, 10) for the day-based
--- queries (/api/days, /api/day/<date>/routes, /api/events?date=).
--- Plain idx_trips_start_time / idx_events_timestamp are NOT used by
--- SQLite when the WHERE clause wraps the column in substr() — verified
--- via EXPLAIN QUERY PLAN. Without these expression indexes the day
--- view degrades to a full table scan on every nav, which is unusable
--- on a Pi Zero 2 W with a few thousand trips.
-CREATE INDEX IF NOT EXISTS idx_trips_day
-    ON trips(substr(start_time, 1, 10));
-CREATE INDEX IF NOT EXISTS idx_events_day
-    ON detected_events(substr(timestamp, 1, 10));
--- Worker pick-next index: partial index over only unclaimed, ready-to-run
--- rows. Keeps the atomic-claim subquery O(log n) regardless of queue depth.
-CREATE INDEX IF NOT EXISTS idx_queue_ready
-    ON indexing_queue(priority, enqueued_at)
-    WHERE claimed_by IS NULL;
--- Stale-claim recovery scan: lets the worker quickly find rows whose
--- claim has aged out (>30 min, indicating the previous worker crashed).
-CREATE INDEX IF NOT EXISTS idx_queue_claimed_at
-    ON indexing_queue(claimed_at)
-    WHERE claimed_by IS NOT NULL;
-
--- v10: archive_queue. Producer-only in Phase 2a (issue #76); the worker
--- that drains it lands in Phase 2b. Rows accumulate harmlessly until
--- then. Keyed by ``source_path`` (UNIQUE) so the inotify producer, the
--- 60-s rescan producer, and the boot catch-up scan can all use
--- ``INSERT OR IGNORE`` for cheap idempotent enqueue. ``priority``
--- follows the issue spec: 1=RecentClips (Tesla rotates these out
--- after ~60 min, highest urgency), 2=SentryClips/SavedClips
--- (event clips the user wants ASAP), 3=anything else. ``status``
--- transitions through pending → claimed → copied (terminal) or
--- → source_gone / error / dead_letter (terminal). ``expected_size``
--- and ``expected_mtime`` are captured at enqueue time so the Phase
--- 2b worker can detect "Tesla still writing" by re-stat-ing before
--- the copy.
-CREATE TABLE IF NOT EXISTS archive_queue (
-    id INTEGER PRIMARY KEY,
-    source_path TEXT UNIQUE NOT NULL,
-    dest_path TEXT,
-    priority INTEGER DEFAULT 3,
-    status TEXT DEFAULT 'pending',
-    attempts INTEGER DEFAULT 0,
-    last_error TEXT,
-    enqueued_at TEXT NOT NULL,
-    claimed_at TEXT,
-    claimed_by TEXT,
-    copied_at TEXT,
-    expected_size INTEGER,
-    expected_mtime REAL
-);
--- Worker pick-next index: partial over only ready rows, ordered by
--- priority then mtime (closest-to-TTL first within each priority band).
--- The worker's pick query is ``SELECT ... WHERE status='pending' ORDER
--- BY priority ASC, expected_mtime ASC LIMIT 1``; this index makes it
--- O(log n) regardless of queue depth.
-CREATE INDEX IF NOT EXISTS archive_queue_ready
-    ON archive_queue(status, priority, expected_mtime)
-    WHERE status = 'pending';
-"""
-
-
-def _backup_db(db_path: str, target_version: int) -> Optional[str]:
-    """Make a copy of the DB before a destructive migration.
-
-    Returns the backup path on success, None on failure (migration still proceeds).
-    Old backups beyond ``_BACKUP_RETENTION`` are pruned.
-    """
-    if not os.path.isfile(db_path):
-        return None
-    try:
-        ts = datetime.now().strftime("%Y%m%d-%H%M%S")
-        backup_path = f"{db_path}.bak.v{target_version}.{ts}"
-        shutil.copy2(db_path, backup_path)
-        logger.info("Backed up geo-index DB to %s", backup_path)
-
-        # Prune older backups
-        backups = sorted(
-            f for f in os.listdir(os.path.dirname(db_path) or '.')
-            if f.startswith(os.path.basename(db_path) + '.bak.')
-        )
-        if len(backups) > _BACKUP_RETENTION:
-            for old in backups[:-_BACKUP_RETENTION]:
-                try:
-                    os.remove(os.path.join(os.path.dirname(db_path), old))
-                except OSError:
-                    pass
-        return backup_path
-    except Exception as e:
-        logger.warning("Failed to back up DB before migration: %s", e)
-        return None
-
-
-def _init_db(db_path: str) -> sqlite3.Connection:
-    """Initialize the SQLite database with schema if needed."""
-    os.makedirs(os.path.dirname(db_path), exist_ok=True)
-
-    conn = sqlite3.connect(db_path, timeout=15)
-    conn.row_factory = sqlite3.Row
-    # Tuned for Pi Zero 2 W (512 MB RAM) where mmap exhaustion under
-    # concurrent indexer + web load was producing spurious "disk I/O error"
-    # responses from SQLite. The combination of a small per-connection page
-    # cache, no file mmap, capped WAL size, and frequent autocheckpoint
-    # keeps each connection's memory footprint bounded so we never run out
-    # of address space when many waitress threads open connections in
-    # parallel during a heavy indexer run.
-    conn.execute("PRAGMA journal_mode=WAL")
-    conn.execute("PRAGMA synchronous=NORMAL")
-    conn.execute("PRAGMA busy_timeout=15000")
-    conn.execute("PRAGMA cache_size=-2000")        # 2 MB per connection
-    conn.execute("PRAGMA mmap_size=0")             # disable file-mmap (saves vmem)
-    conn.execute("PRAGMA temp_store=MEMORY")       # avoid temp files on slow SD
-    conn.execute("PRAGMA journal_size_limit=4194304")   # cap WAL at 4 MB
-    conn.execute("PRAGMA wal_autocheckpoint=200")  # checkpoint every ~800 KB
-    conn.execute("PRAGMA foreign_keys=ON")
-
-    # Check schema version. Older code used INSERT OR REPLACE on a PRIMARY KEY
-    # column, which actually added a new row each time, so older DBs may have
-    # multiple rows. Use MAX() to read the effective version.
-    try:
-        row = conn.execute("SELECT MAX(version) AS v FROM schema_version").fetchone()
-        current = row['v'] if row and row['v'] is not None else 0
-    except sqlite3.OperationalError:
-        current = 0
-
-    if current < _SCHEMA_VERSION:
-        # Backup before any destructive migration (only when an existing DB
-        # is being upgraded, not on first install)
-        if current > 0:
-            _backup_db(db_path, _SCHEMA_VERSION)
-
-        conn.executescript(_SCHEMA_SQL)
-        # Migrations for existing databases
-        if current < 2:
-            # v2: add blinker columns to waypoints
-            for col in ('blinker_on_left', 'blinker_on_right'):
-                try:
-                    conn.execute(f"ALTER TABLE waypoints ADD COLUMN {col} INTEGER DEFAULT 0")
-                except sqlite3.OperationalError:
-                    pass  # Column already exists
-        if current > 0 and current < 3:
-            # v3: clean up duplicate trips/waypoints from earlier indexer bugs.
-            # Wrapped in a savepoint so a failure during migration doesn't leave
-            # the schema_version bumped without the data fixes applied.
-            try:
-                conn.execute("SAVEPOINT migrate_v3")
-                _migrate_v2_to_v3(conn)
-                conn.execute("RELEASE SAVEPOINT migrate_v3")
-            except Exception as e:
-                conn.execute("ROLLBACK TO SAVEPOINT migrate_v3")
-                conn.execute("RELEASE SAVEPOINT migrate_v3")
-                logger.error("Migration v2->v3 failed, leaving schema at v2: %s", e)
-                conn.commit()
-                return conn  # Skip schema_version bump so it retries next startup
-        if current > 0 and current < 4:
-            # v4: re-evaluate Sentry/Saved clips with Tesla's event.json
-            # (which has accurate GPS) instead of the prior nearest-waypoint
-            # guess. We do this by clearing their indexed_files rows so the
-            # next indexer run re-processes them through the new code path.
-            try:
-                conn.execute("SAVEPOINT migrate_v4")
-                _migrate_v3_to_v4(conn)
-                conn.execute("RELEASE SAVEPOINT migrate_v4")
-            except Exception as e:
-                conn.execute("ROLLBACK TO SAVEPOINT migrate_v4")
-                conn.execute("RELEASE SAVEPOINT migrate_v4")
-                logger.error("Migration v3->v4 failed, leaving schema at v3: %s", e)
-                conn.commit()
-                return conn
-        if current > 0 and current < 9:
-            # v9: one-shot repair pass for trips that were fragmented
-            # by the matching-SQL boundary bug fixed in this version.
-            # The bug: ``ORDER BY ABS(new_start - existing.start)`` plus
-            # the float-imprecise ``(julianday(...)-julianday(...))*86400``
-            # condition caused phantom-fragmented trips when files
-            # arrived out-of-order during indexer pauses (e.g., archive-
-            # lock starvation incident May 2026 — McDonald's drive split
-            # into 6 trips). The runtime ``_merge_adjacent_trips_for``
-            # added in this version prevents future fragmentation, but
-            # only sweeps the just-touched anchor's neighbourhood, so
-            # bad data already in the table will linger unless a future
-            # clip happens to bridge it. This one-shot global merge
-            # repairs the existing damage.
-            try:
-                conn.execute("SAVEPOINT migrate_v9")
-                merged = _merge_all_adjacent_trip_pairs(
-                    conn, _TRIP_GAP_MINUTES_DEFAULT * 60,
-                )
-                conn.execute("RELEASE SAVEPOINT migrate_v9")
-                if merged:
-                    logger.info(
-                        "Migration v8->v9: merged %d phantom-fragmented "
-                        "trip pairs", merged,
-                    )
-            except Exception as e:
-                conn.execute("ROLLBACK TO SAVEPOINT migrate_v9")
-                conn.execute("RELEASE SAVEPOINT migrate_v9")
-                logger.error(
-                    "Migration v8->v9 failed, leaving schema at v8: %s", e,
-                )
-                conn.commit()
-                return conn
-        # v5: covering index ``idx_waypoints_trip_video`` for the
-        # ``/api/trips`` page-load N+1 fix. The index is created by the
-        # ``executescript(_SCHEMA_SQL)`` call above (CREATE INDEX IF NOT
-        # EXISTS), so no separate data migration is needed — the schema
-        # version bump is the trigger.
-        # v6: ``indexing_queue`` table for the queue-based indexer
-        # redesign. Created by the executescript call; no data migration
-        # because there's nothing in the queue on the first upgrade — the
-        # boot catch-up scan will repopulate from indexed_files diff.
-        # v7: ``idx_trips_start_time`` for callers that range-scan
-        # by ``start_time`` directly (e.g. /api/trips with date_from
-        # bounds). Created by the executescript above.
-        # v8: expression indexes ``idx_trips_day`` and
-        # ``idx_events_day`` on ``substr(<ts>, 1, 10)`` so the day-
-        # based map queries can avoid full scans. Plain timestamp
-        # indexes do NOT cover ``substr(col, 1, 10) = ?``. The
-        # expression indexes are created by the executescript above
-        # (CREATE INDEX IF NOT EXISTS is idempotent); no data
-        # migration required.
-        # v10: ``archive_queue`` table + ``archive_queue_ready``
-        # partial index for the Phase 2a producers (issue #76).
-        # Producer-only at this version — the boot catch-up scan,
-        # 60-s rescan, and inotify file watcher all enqueue rows but
-        # nothing drains them yet. The Phase 2b worker (separate PR)
-        # will be the consumer. Created by the executescript above
-        # (CREATE TABLE IF NOT EXISTS / CREATE INDEX IF NOT EXISTS
-        # are idempotent); no data migration required.
-        conn.execute("DELETE FROM schema_version")
-        conn.execute(
-            "INSERT INTO schema_version (version) VALUES (?)",
-            (_SCHEMA_VERSION,)
-        )
-        conn.commit()
-        logger.info("Geo-index database initialized (v%d) at %s", _SCHEMA_VERSION, db_path)
-
-    return conn
-
-
-# ---------------------------------------------------------------------------
-# Migrations
-# ---------------------------------------------------------------------------
-
-def _migrate_v2_to_v3(conn: sqlite3.Connection) -> None:
-    """Clean up duplicate trips and waypoints from earlier indexer bugs.
-
-    Earlier versions of the indexer:
-      * Created separate trips for the same physical drive when the videos
-        were ingested from different source folders (RecentClips vs ArchivedClips).
-      * Stored duplicate waypoints with the same ``(timestamp, lat, lon)`` but
-        different ``video_path`` strings (one per copy of the video).
-      * Recorded ``source_folder='..'`` for ArchivedClips because of a path
-        normalization bug.
-
-    This one-time migration:
-      1. Repairs ``source_folder='..'`` rows by inferring from waypoint paths.
-      2. Merges trips whose time windows overlap or are within
-         ``_TRIP_GAP_MINUTES_DEFAULT`` minutes of each other (regardless of
-         source_folder).
-      3. Dedupes waypoints within each trip by ``(timestamp, lat, lon)``,
-         preferring the row whose ``video_path`` references ArchivedClips
-         (most durable storage).
-      4. Recomputes ``start_time``, ``end_time``, start/end coords,
-         ``distance_km`` and ``duration_seconds`` for every trip; deletes
-         trips left with no waypoints.
-    """
-    gap_seconds = _TRIP_GAP_MINUTES_DEFAULT * 60
-    log_parts: List[str] = []
-
-    # --- Phase 1: source_folder='..' ---
-    bad = conn.execute(
-        "SELECT id FROM trips WHERE source_folder = '..' OR source_folder LIKE '..%'"
-    ).fetchall()
-    fixed_src = 0
-    for r in bad:
-        wp = conn.execute(
-            "SELECT video_path FROM waypoints "
-            "WHERE trip_id = ? AND video_path IS NOT NULL ORDER BY id LIMIT 1",
-            (r['id'],),
-        ).fetchone()
-        if wp and wp['video_path']:
-            vp = wp['video_path'].replace('\\', '/')
-            if 'ArchivedClips' in vp:
-                folder = 'ArchivedClips'
-            elif '/' in vp:
-                folder = vp.split('/')[0]
-            else:
-                folder = 'Unknown'
-            conn.execute(
-                "UPDATE trips SET source_folder = ? WHERE id = ?",
-                (folder, r['id']),
-            )
-            fixed_src += 1
-    log_parts.append(f"fixed {fixed_src} '..' source_folder rows")
-
-    # --- Phase 2: merge overlapping/close trips ---
-    # Repeatedly find any pair of trips whose windows are within gap_seconds
-    # of each other (in either direction) and merge the higher-id into the lower.
-    merged = _merge_all_adjacent_trip_pairs(conn, gap_seconds)
-    log_parts.append(f"merged {merged} overlapping trip pairs")
-
-    # --- Phase 3: dedupe waypoints within a trip ---
-    dups = conn.execute(
-        """SELECT trip_id, timestamp, lat, lon, COUNT(*) AS cnt
-           FROM waypoints
-           WHERE trip_id IS NOT NULL
-           GROUP BY trip_id, timestamp, lat, lon
-           HAVING COUNT(*) > 1"""
-    ).fetchall()
-    deduped = 0
-    for d in dups:
-        ids = conn.execute(
-            """SELECT id, video_path FROM waypoints
-               WHERE trip_id = ? AND timestamp = ? AND lat = ? AND lon = ?
-               ORDER BY
-                 CASE WHEN video_path LIKE '%ArchivedClips%' THEN 0 ELSE 1 END,
-                 id""",
-            (d['trip_id'], d['timestamp'], d['lat'], d['lon']),
-        ).fetchall()
-        # Keep the first (durable / lowest id), delete the rest
-        drop_ids = [(r['id'],) for r in ids[1:]]
-        if drop_ids:
-            conn.executemany("DELETE FROM waypoints WHERE id = ?", drop_ids)
-            deduped += len(drop_ids)
-    log_parts.append(f"deduped {deduped} duplicate waypoints")
-
-    # --- Phase 4: recompute trip stats; drop empty trips ---
-    # Distance is computed per video file (in frame/id order) and summed,
-    # because Tesla videos can overlap in time (e.g. when a saved clip is
-    # triggered alongside RecentClips). Sorting all waypoints globally by
-    # timestamp would interleave overlapping recordings and produce huge
-    # phantom jumps. start_time/end_time still come from min/max timestamp.
-    trips = conn.execute("SELECT id FROM trips").fetchall()
-    recomputed = 0
-    dropped_empty = 0
-    for t in trips:
-        bounds = conn.execute(
-            "SELECT MIN(timestamp) AS first_ts, MAX(timestamp) AS last_ts "
-            "FROM waypoints WHERE trip_id = ?",
-            (t['id'],),
-        ).fetchone()
-        if not bounds or not bounds['first_ts']:
-            conn.execute("DELETE FROM trips WHERE id = ?", (t['id'],))
-            dropped_empty += 1
-            continue
-        first_ts, last_ts = bounds['first_ts'], bounds['last_ts']
-        first_row = conn.execute(
-            "SELECT lat, lon FROM waypoints WHERE trip_id = ? "
-            "AND timestamp = ? ORDER BY id LIMIT 1",
-            (t['id'], first_ts),
-        ).fetchone()
-        last_row = conn.execute(
-            "SELECT lat, lon FROM waypoints WHERE trip_id = ? "
-            "AND timestamp = ? ORDER BY id DESC LIMIT 1",
-            (t['id'], last_ts),
-        ).fetchone()
-        # Distance summed per video file
-        total_dist = 0.0
-        videos = conn.execute(
-            "SELECT DISTINCT video_path FROM waypoints "
-            "WHERE trip_id = ? AND video_path IS NOT NULL",
-            (t['id'],),
-        ).fetchall()
-        for v in videos:
-            wps = conn.execute(
-                "SELECT lat, lon FROM waypoints "
-                "WHERE trip_id = ? AND video_path = ? ORDER BY id",
-                (t['id'], v['video_path']),
-            ).fetchall()
-            for j in range(1, len(wps)):
-                total_dist += _haversine_km(
-                    wps[j-1]['lat'], wps[j-1]['lon'],
-                    wps[j]['lat'], wps[j]['lon'],
-                )
-        try:
-            dur = max(0, int((
-                datetime.fromisoformat(last_ts)
-                - datetime.fromisoformat(first_ts)
-            ).total_seconds()))
-        except (ValueError, TypeError):
-            dur = 0
-        conn.execute(
-            """UPDATE trips SET
-               start_time = ?, end_time = ?,
-               start_lat = ?, start_lon = ?,
-               end_lat = ?, end_lon = ?,
-               distance_km = ?, duration_seconds = ?
-               WHERE id = ?""",
-            (first_ts, last_ts,
-             first_row['lat'] if first_row else None,
-             first_row['lon'] if first_row else None,
-             last_row['lat'] if last_row else None,
-             last_row['lon'] if last_row else None,
-             total_dist, dur, t['id']),
-        )
-        recomputed += 1
-    log_parts.append(
-        f"recomputed stats for {recomputed} trips; dropped {dropped_empty} empty"
-    )
-
-    logger.info("Migration v2->v3: %s", "; ".join(log_parts))
+#
+# The schema DDL, version constants, backup helper, ``_init_db`` connection
+# factory, and the v2/v3/v4 migrations now live in
+# ``services.mapping_migrations``. ``_init_db``, ``_backup_db`` and
+# ``_SCHEMA_VERSION`` are re-exported from this module (top-level import
+# below) so the (very many) existing internal call sites and test imports
+# continue to work unchanged. New code should import from
+# ``services.mapping_migrations`` directly.
 
 
 # Default trip gap, also used by the migration. Kept here so the migration
@@ -857,51 +374,6 @@ def _merge_all_adjacent_trip_pairs(conn: sqlite3.Connection,
 
 # ---------------------------------------------------------------------------
 # Event Detection Rules
-# ---------------------------------------------------------------------------
-
-def _migrate_v3_to_v4(conn: sqlite3.Connection) -> None:
-    """Re-evaluate Sentry/Saved clips with Tesla's event.json.
-
-    Earlier versions inferred Sentry/Saved event locations from the
-    nearest waypoint, which was inaccurate (often pointed at a different
-    physical location). Tesla actually writes a precise event.json with
-    est_lat/est_lon in each event folder.
-
-    To pick this up for clips already in the database, we delete:
-      1. The existing inferred-location detected_events rows
-         (they have metadata.inferred_location=true), and
-      2. The indexed_files rows for SavedClips/SentryClips clips with
-         zero waypoints — so the next indexer run re-processes them
-         through the new event.json-aware code path.
-
-    Driving clips (those with waypoints) are left untouched.
-    """
-    # Drop old inferred events so they get recreated from event.json
-    cur = conn.execute(
-        "DELETE FROM detected_events "
-        "WHERE event_type IN ('saved', 'sentry') "
-        "AND metadata IS NOT NULL "
-        "AND (metadata LIKE '%inferred_location%' "
-        "     OR metadata LIKE '%nearest_waypoint%')"
-    )
-    deleted_events = cur.rowcount
-    logger.info("v3->v4: cleared %d stale inferred events", deleted_events)
-
-    # Clear indexed_files rows for SavedClips/SentryClips zero-waypoint
-    # entries so they get re-indexed with event.json reading
-    cur = conn.execute(
-        "DELETE FROM indexed_files "
-        "WHERE waypoint_count = 0 "
-        "AND (file_path LIKE '%/SavedClips/%' "
-        "     OR file_path LIKE '%/SentryClips/%' "
-        "     OR file_path LIKE '%\\SavedClips\\%' "
-        "     OR file_path LIKE '%\\SentryClips\\%')"
-    )
-    cleared_files = cur.rowcount
-    logger.info("v3->v4: cleared %d Sentry/Saved indexed_files entries for re-processing",
-                cleared_files)
-
-
 # ---------------------------------------------------------------------------
 
 # Default thresholds (can be overridden via config.yaml mapping.event_detection)
